Split ninja_test.cc into state_test.cc and disk_interface_test.cc
