+++ configure.py
@@ -172,9 +172,17 @@ else:
     test_libs = libs + ['-lgtest_main', '-lgtest']
 
 objs = []
-for name in ['build_test', 'build_log_test', 'disk_interface_test',
-             'eval_env_test', 'graph_test', 'ninja_test', 'parsers_test',
-             'subprocess_test', 'util_test', 'clean_test', 'test']:
+for name in ['build_log_test',
+             'build_test',
+             'clean_test',
+             'disk_interface_test',
+             'eval_env_test',
+             'graph_test',
+             'parsers_test',
+             'state_test',
+             'subprocess_test',
+             'test',
+             'util_test']:
     objs += cxx(name, variables=[('cflags', test_cflags)])
 
 if platform != 'mingw':
